requirements.txt
numpy
opencv-python
pillow
matplotlib
tqdm
torch
torchvision
python-chess
scikit-image
mediapipe